requirements.txt
cloudscraper
os
bs4==0.0.1
by==0.0.4
cfscrape==2.1.1
chromedriver-autoinstaller==0.6.3
chromedrivermanager==0.0.1
discord-webhook==1.1.0
discordwebhook==1.0.3
openpyxl==3.1.2
pandas==2.0.3
pip==20.2.3
pipdeptree==2.13.2
pylint==2.6.0
pyOpenSSL==23.3.0
pyowm==3.3.0
python-dotenv==1.0.0
pywhatkit==5.4
selenium==4.0.0
service==0.6.0
webdriver-manager==3.5.2
xlwt==1.3.0
